+++ src/components/ids-data-grid/ids-data-grid-column.ts
@@ -94,6 +94,10 @@ export interface IdsDataGridColumn {
   field?: string;
   /** The subsitute text to use (for hyperlink and some formatters) */
   text?: string;
+  /** Max value of a range */
+  max?: number;
+  /** Mininum value of a range */
+  min?: number;
   /** Allow column sorting */
   sortable?: boolean;
   /** Allow column resizing */
@@ -118,6 +122,8 @@ export interface IdsDataGridColumn {
   click?: (rowData: Record<string, unknown>, columnData: IdsDataGridColumn, event: MouseEvent) => void;
   /** Get the color dynamically from a function or as text */
   color?: string | ((row: number, value: any, column: IdsDataGridColumn, index: Record<string, any>) => string | undefined);
+  /** Get the size dynamically from a function or as text */
+  size?: string | ((row: number, value: any, column: IdsDataGridColumn, index: Record<string, any>) => string | undefined);
   /** Options to pass to the formatter */
   formatOptions?: IdsDataGridColumnFormatOptions;
   /** Column Filter to use */
@@ -205,7 +211,7 @@ export interface IdsDataGridColumn {
   cellSelectedCssPart?: string | ((rowIndex: number, cellIndex: number) => string);
   /** Setup an editor */
   editor?: {
-    type: 'input' | 'date' | 'time' | 'checkbox' | 'dropdown',
+    type: 'input' | 'date' | 'time' | 'checkbox' | 'dropdown' | 'datepicker' | 'timepicker',
     inline?: boolean,
     editor?: IdsDataGridEditor,
     editorSettings?: Record<string, unknown>