+++ src/components/ids-time-picker/ids-time-picker.ts
@@ -116,7 +116,7 @@ export default class IdsTimePicker extends Base {
    * List of available color variants for this component
    * @returns {Array<string>}
    */
-  colorVariants = ['alternate-formatter'];
+  colorVariants = ['alternate-formatter', 'borderless', 'in-cell'];
 
   /**
    * Push color variant to the trigger-field element
@@ -267,6 +267,7 @@ export default class IdsTimePicker extends Base {
    */
   #attachKeyboardListeners(): IdsTimePicker {
     this.listen(['ArrowDown', 'Escape', 'Backspace'], this, (e: KeyboardEvent) => {
+      e.stopPropagation();
       if (e.key === 'ArrowDown') {
         this.picker?.show();
       } else if (e.key === 'Escape' || e.key === 'Backspace') {